pyproject.toml
[project]
name = "mcp-thisweek"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
]
